- разбил структуры по пакетам - использовал приватные интерфейсы
