+++ pkg/builder/cleanUpTheRoom/cleanUpTheRoom.go
@@ -1,34 +1,39 @@
-package builder
+package cleanUpTheRoom
 
-import "fmt"
+import (
+	"fmt"
+)
 
 // roomCleaning implements child interface
 type roomCleaning struct {
-	order *order
+	order order
 }
 
 // MakeTheBed makes the bed
 func (r *roomCleaning) MakeTheBed(direction string) {
-	r.order.result += fmt.Sprintf("Made the bed %v\n", direction)
+	result := fmt.Sprintf("Made the bed %v\n", direction)
+	r.order.AddResult(result)
 }
 
 // CollectToys collects the toys
 func (r *roomCleaning) CollectToys(direction string) {
-	r.order.result += fmt.Sprintf("Collected %v toys\n", direction)
+	result := fmt.Sprintf("Collected %v toys\n", direction)
+	r.order.AddResult(result)
 }
 
 // VacuumTheFloor cleans the floor using vacuum cleaner
 func (r *roomCleaning) VacuumTheFloor(direction string) {
-	r.order.result += fmt.Sprintf("Vacuumed the floor %v\n", direction)
+	result := fmt.Sprintf("Vacuumed the floor %v\n", direction)
+	r.order.AddResult(result)
 }
 
-// Call implementation
-func (r *roomCleaning) Call() {
-	fmt.Println("...starting of room cleaning")
+// order provides an order interface
+type order interface {
+	AddResult(string)
 }
 
 // NewRoomCleaning creates roomCleaning
-func NewRoomCleaning(order *order) *roomCleaning {
+func NewRoomCleaning(order order) *roomCleaning {
 	return &roomCleaning{
 		order: order,
 	}
